🌱 Split auth/LocalLoginForm.tsx and onboarding/Tour.tsx
Fixes #21793

Extract useLocalLogin hook + LocalLoginForm.parts components from LocalLoginForm.tsx; extract getTooltipPosition + positioning types/constants from Tour.tsx into Tour.parts.tsx. Both source files now under 350 lines with identical runtime behaviour.
